nvme-discover: Retry discovery log if the generation counter changes
Git-commit: 3cbcd165b470a1b86e323e2b256cde8f05a9b5dd
References: bsc#1090568

If the generation counter changes we need to validate if the number
of records has changed, too.
If so we need to retry retrieving the discovery log to the most recent
values. The retry will be terminated after MAX_DISC_RETRIES (currently
set to 10) to avoid infinite recursion.

Signed-off-by: Hannes Reinecke <hare@suse.com>