nvme-ioctl: retrieve log pages in 4k chunks
Git-commit: 465a4d5ab209bb20ea4b3ba808438747773edd2b
References: bsc#1105314

Some log pages increase with the scale of the subsystem, so the
overall size might be larger than the MDTS value for that controller.
In order to avoid having to pass in the MDTS value we should restrict
the transfer size to 4k and retrieve the log page in 4k chunks.

Signed-off-by: Hannes Reinecke <hare@suse.com>