nvme-discover: Re-check generation counter after log page transfer
Git-commit: 36efbbbff9f0e5d854f0454e393012ae23c46646
References: bsc#1076004

The log page transfer might have been split up in several chunks, so
it might happen that the generation counter changed inbetween these
transfer. Hence we need to re-fetch the header again to figure out
if there had been changes to the generation counter; if so we need
to fetch the entire page again.

Signed-off-by: Hannes Reinecke <hare@suse.com>