nvme-cli: fix fc trtype string
Commit 55b4ec12ffe4
("fabrics: use trtype_str when parsing a discovery log entry")
changed to use the trtypes string names in a generic fashion for
the "transport=?" string passed in the connect string to the kernel.

For FC, this meant the hardcoding of "transport=fc" was removed and
the generic code would now create "transport=fibre-channel", which
doesn't match any transport name anymore once the string is attempted
to be matched in the kernel.

Change the string used by for trtype for FC to use "fc" to restore
operation.

Signed-off-by: James Smart <jsmart2021@gmail.com>