Change nvmefc-connect@.service to type `simple`
The idea is that udevd calls systemd for the nvme connect-all calls so
that it doesn't have to wait for it to return. But systemd waits for
services of type `oneshot` when called with systemctl start until the
process specified in the service file terminates.

Instead use type `simple`: systemd will return immediately after
starting the service, regardless of how long the nvme connect-all call
takes.

Signed-off-by: Simon Schricker <sschricker@suse.de>